requirements.txt
datasets==2.12.0
Flask==2.3.1
openai==0.27.5
pinecone_client==2.2.1
python-dotenv==1.0.0
tqdm==4.65.0
